修复了一些Bug

+++ pojo/src/main/java/com/forfries/dto/ScreeningHallPageDTO.java
@@ -6,6 +6,6 @@ import java.io.Serializable;
 
 @Data
 public class ScreeningHallPageDTO extends PageDTO implements Serializable {
-    private Integer cinemaId;
+    private Long cinemaId;
     private String name;
 }
